requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# AI/ML libraries
mem0ai>=0.1.0
openai>=1.3.0
langchain>=0.1.0

# Database clients
qdrant-client>=1.7.0
neo4j>=5.15.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
python-multipart>=0.0.6

# Optional: Monitoring
langfuse>=2.0.0

# Development
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
